Periodic sonar run
Small cleanups as reported by sonarqube.

+++ community/kernel/src/main/java/org/neo4j/kernel/impl/util/ValueUtils.java
@@ -382,7 +382,7 @@ public final class ValueUtils
         }
         if ( object instanceof Integer )
         {
-            return Values.longValue( (long) object );
+            return Values.longValue( (int) object );
         }
 
         throw new IllegalArgumentException(
